+++ libs/radcore/src/radfile/common/filesystem.cpp
@@ -278,7 +278,7 @@ void radFileSystem::ProcessFileName
     const char* drive = strchr( pFileName, ':' );
     if( simpleName==false && drive != NULL )
     {
-        length = (unsigned int) drive - (unsigned int) pFileName + 1;
+        length = (uintptr_t) drive - (uintptr_t) pFileName + 1;
         strncpy( driveSpec, pFileName, length );
         driveSpec[ length ] = '\0';
         fname = &pFileName[ length ];
